requirements.txt
pip==22.3.1
setuptools==65.5.1
wheel==0.38.4
chainlit~=0.5.1
python-dotenv~=1.0.0
langchain~=0.0.228